backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
anthropic==0.7.8
httpx==0.25.2
python-multipart==0.0.6
fastapi-cors